Improve formatting of fields/alternatives
If a record field is a `Some` or a completion (or both), this formats them
on the same line as the key to improve the compactness of the formatted
output.

Note that this change applies to record types and union types, too, but in
practice we only expect to see `Some` and record completions in record fields.