Relocate dhall-lang.org website (#1642)
* Relocate dhall-lang.org website

This is a matching change to https://github.com/dhall-lang/dhall-lang/pull/892

Now that the code for building the website lives in the `dhall-lang`
repository we no longer need to keep the code here.  This repository is now
only responsible for building the `dhall-try` package which provides the
GHCJS code to power the live demo.

* Document how to test GHCJS changes

... based on an inquiry from @sjakobi