Drop support for GHC 7.10.3
The immediate motivation for doing this is so that we don't need to
introduce yet another GHC 7.10.3 workaround for
https://github.com/dhall-lang/dhall-haskell/pull/1640

More generally, the main reason we were still supporting GHC 7.10.3
was for Eta compatibility, but Eta appears to no longer be actively
developed.  Given that Eta seems unlikely to support GHC 8 and we
don't plan on supporting GHC 7.10.3 in perpetuity it's probably
time to drop support.

We could probably drop support for additional GHC versions, but this
is just the smallest change necessary to unblock #1640.