Support template haskell for multiple datatypes (#1664)
Inspired by this discussion: https://discourse.dhall-lang.org/t/makehaskelltypefromunion-and-fields-with-sum-types/150

This adds a new `makeHaskellTypes` utility which can generate multiple
types which can refer to each other.  The trick is that the type
compares nested Dhall types against top-level ones and replaces them
with Haskell datatype references when the types are judgmentally
equivalent.