Fix `FromDhall` instance for `Fix`
The root cause of the test failure was that the `extract` function's
behavior didn't match the `expected` type.  Specifically, the
`extract` function was written to work on an α-normalized expression,
whereas the `expected` type was not α-normalized.

There are two ways to fix this problem:

* Have both the `extract` and `expected` both expect the α-normalized
  term / type

* Have them both expect the non-α-normalized type that begins with
  ∀-bound `result` and `Make` variables

I went with the latter approach.  Specifically, the `expected` type
remains the same, whereas the extract function now renames variables
within the extract expression before substitution to match the
`expected` type.

The reason we didn't catch this failure before this change is because
the `union` `Decoder` is more strict than the previous derived
`FromDhall` instance for unions.  Specifically, the `union` `Decoder`'
`extract` function checks (via a `guard`) that the union type literal
used by the union constructor matches the `expected` union type,
whereas previously no such check was performed.  This check caught a
(minor) bug in the `FromDhall` instance for `Fix` which this change
fixes.