Add support for converting union types to directory trees.
This only works for alternatives with suitable non-nullary constructors. For
example, running `dhall to-directory-tree` on this expression:

    let UnionType : Type = < Left : Text | Right : Text >
    in { foo = UnionType.Left "one", bar = UnionType.Right "two" }

will result in the output directory containing two regular files, `foo` and
`bar`, with the contents "one" and "two", respectively.

The following union type, however, cannot be converted to a directory tree:

    let UnionType2 : Type = < Left : Text | Middle | Right : Text>

because the nullary `Middle` constructor has no associated value.

This change is of particular use when dealing with union types whose
constructors take values of differently structured types:

    let Thinkpad : Type = { series : Text, model : Text }
    let MacBookPro : Type = { year : Text }

    let Laptop : Type = < TP : Thinkpad | MBP : MacBookPro >

    let flywheel : Thinkpad = { series = "T", model = "460" }
    let splinch : MacBookPro = { year = "2007" }

    let myLaptops = {
        flywheel = Laptop.TP flywheel,
        splinch = Laptop.MBP splinch
    }

    in myLaptops

With this change, the above expression can be converted into the following
directory tree:

    <output dir>
    |- flywheel
    |  |- series
    |  \- model
    \- splinch
       \- year

Previously, this would have required creating a record structure which had an
`Optional Text` field for every possible member, something like:

    let Laptop2 : Type = { series : Optional Text, model : Optional Text
                         , year : Optional Text }

which has the issue that it would be possible to introduce an invalid instance
of the `Laptop2` type.