nix/shared.nix: take cabal2nix from pkgs, to not compile statically
cabal2nix is a build dependency, thus it’s not required that we build
it statically.

This strips a huge amount out of the static build tree, because
cabal2nix depends on a *lot* of libraries, plus it references most
build tools in existence, which would *also* have to be built
statically.