Format documents with leading `---`
The context for this is: https://github.com/dhall-lang/dhall-haskell/issues/1864

Not only is this legal, but I believe it's actually more consistent with
how `---` was intended to be used (to separate each YAML document from
its preceding header).  The YAML specification also uses `---` in the
same way for its document-related examples.