Move RecordEncoder documentation
The documentation for how to use the Divisible instance to create a
RecordEncoder ended up in the "Miscellaneous" section. This made it very
hard to find, after some searching I ended up in the repository and
found out the documentation *was* included on the hackage page I was
trying to find it on.

I moved it to document the definition of the RecordEncoder newtype to
bring it in line with the RecordDecoder documentation.