Format documents with leading `---` (#1865)
* Format documents with leading `---`

The context for this is: https://github.com/dhall-lang/dhall-haskell/issues/1864

Not only is this legal, but I believe it's actually more consistent with
how `---` was intended to be used (to separate each YAML document from
its preceding header).  The YAML specification also uses `---` in the
same way for its document-related examples.

* Add test for single-document case

... as requested by @sjakobi

* Update documentation for `--documents` option

... as requested by @sjakobi

* Add test case for `--documents` without a list

... based on a discussion with @sjakobi

* Build against `aeson-yaml-1.1.0`

* Fix `aeson-yaml` version in `stack.yaml`

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>