Add new `dhall schema` subcommand
Fixes https://github.com/dhall-lang/dhall-haskell/issues/1796

This adds a new `dhall schema` subcommand that one can use to simplify
a Dhall expression using a supplied record of schemas.  Any time a
subexpression matches the type of a schema it will be replaced with the
corresponding schema.

If more than one schema matches the supplied type then the first match
(in alphabetical order) is used.