Greatly improve efficiency of generators
CI was hanging on the property test due to filtering out the overwhelming
majority of generated `File`s using `suchThat`.  The issue was that the
vast majority of characters were not within the valid range, so the
chance of generating a valid `File` was vanishingly small.

The fix is to rewrite the generation logic so that the generated
`File`s are correct by construction without having to filter them down.