Fix `Arbitrary` instances for `File`s (Take 2) (#1937)
* Fix `Arbitrary` instances for `File`s (Take 2)

Fixes https://github.com/dhall-lang/dhall-haskell/issues/1934

This also addresses the following comment:

https://github.com/dhall-lang/dhall-haskell/pull/1933#issuecomment-661337567

* Greatly improve efficiency of generators

CI was hanging on the property test due to filtering out the overwhelming
majority of generated `File`s using `suchThat`.  The issue was that the
vast majority of characters were not within the valid range, so the
chance of generating a valid `File` was vanishingly small.

The fix is to rewrite the generation logic so that the generated
`File`s are correct by construction without having to filter them down.

* Support older versions of QuickCheck

... as caught by @sjakobi

* Fix bounds on QuickCheck

Co-authored-by: Simon Jakobi <simon.jakobi@gmail.com>

Co-authored-by: Simon Jakobi <simon.jakobi@gmail.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>