Add trailing character to Text bullet point (#1948)
* Add trailing character to Text bullet point

The bullet point for `Text` at the top of the Dhall.JSON module
documentation was rendered as a code block because there were no
characters following the inline literal.

Added a trailing character so the literal is properly interpreted as
inline rather than block. Selected the Braille Pattern Blank U+2800
because it's a blank character that is not recognized as whitespace.
Alternatives include appending an s, `@Text@s`, or adding a neutral word
as in `@Text@ values`.

* Switch from invisible character to neutral word

Co-authored-by: Simon Jakobi <simon.jakobi@gmail.com>

Co-authored-by: Simon Jakobi <simon.jakobi@gmail.com>