Add dhall-openapi (#1946)
The context for this is:

https://discourse.dhall-lang.org/t/expense-proposal-turn-dhall-kubernetes-generator-into-a-standalone-openapi-to-dhall-project/278

We would like to generalize the `dhall-kubernetes-generator` package from
the `dhall-kubernetes` project into a `dhall-openapi` package (residing
within this repository).  This change imports the package mostly "as is" as the
first step of importing the project, meaning that many of the Kubernetes-specific
features still remain and need to be generalized.