Fix default `PersistentVolumeClaim` (#2053)
… as caught by @PierreR in https://github.com/dhall-lang/dhall-haskell/pull/2019#issuecomment-694751778

The issue was that the logic for generating the default for a type had a
special case for the `apiVersion` / `kind` fields, which was overriding
the correct `None` defaults.  This change fixes that by switching the
priority so that if the `None` defaults are present then they take
precedence over the hard-coded defaults.