Improve cache warning messages
The immediate motivation for this fix is to ensure that
command-line utilities other than `dhall` correctly warn about
issues with cache directories.

The main change along those lines is to move the cache check from
the `Dhall.Main` module to checking on the fly as cache files are
read/written.  However, in order to avoid going back to too many
cache warnings this also includes logic to avoid warning more than
once by tracking a `cacheWarning` state `Bool`.

This also includes a small improvement to the permissions error
message to better render the directory permissions.