Fix `dhall directory --file …` (#2092)
This is to fix an issue reported in:

https://github.com/dhall-lang/dhall-haskell/issues/2091#issuecomment-725707959

The issue is that the `--file` option of the `directory` subcommand
didn't work if the specified `--file` was not in the top-level of the
package.  This is because the base directory for import resolution
was being improperly set to the directory supplied to `dhall-nixpkgs`,
when it should have been the parent directory of the file being imported.

This also includes a small fix to the filename displayed in error
messages (which were incorrectly displaying the parent directory name).

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>