Only put models with GVK in types union
The types unions main/only purpose was to allow creating lists of objects to
output to a single yaml document for sending to kubernetes. These
objects must have a group/version and kind. Therefore any objects
without would not be used in this manner and should be removed from the
typesUnion.