Fix tests on Windows
This should hopefully fix the tests on Windows

The main issue was the `B.dhall` test files using the wrong
line-ending, which is fixed by telling `git` to check them out as if
they are binary files (and therefore preserving the Unix line endings).

The second issue was the use of a non-existent `/tmp` directory on
Windows.