dhall-toml: implement basic toml -> dhall
Conversion from TOML to Dhall, with a required schema file. I added a
comment about it, but I implemented an intermediate representation for
this conversion because the type that tomland provides is not very easy
to travese for two reasons:

1. the objects are dense, that is an object is stored like {a.b.c = 1}
   and not {a = { b = { c = 1 }}}.
2. the main TOML type is split up by the syntax of how the value was
   defined (keys values pairs, tables, array of tables)

I originally tried to put arrays values in a sum type and fitting that
into tomland's PrefixMap, but I was having a hard time reasoning about
it since the axis for arrays and records were in different data types.