Improve robustness of `:let` command (#2297)
* Improve robustness of `:let` command

Fixes https://github.com/dhall-lang/dhall-haskell/issues/2296

The root cause of the above bug was that the old parser for `:let`
commands was too lenient.  In particular, given a command like:

```
:let x : T = e
```

… the old parser would silently ignore the `: T` part of the command.

This change fixes that by adding support for type annotations and
fixing the `:let` command to exactly match the standard parser in
terms of what expressions it permits.

* Use `NamedFieldPuns`

… as suggested by @sjakobi

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>