Modifying field names in generated Haskell types (#2285)
* Dhall.TH: Configurable {From,To}Dhall instances

This commit adds two fields to both constructors of
`Dhall.TH.HaskellType`: One flag to control whether a `FromDhall`
instance will be generated and one to control whether a `ToDhall`
instance will be generated.

* Moved options related to code generation to own data type

* Fixed tests

* Generate {From,To}Dhall instances explicitly

* Implementend `constructorModifier`

This function controls how a Dhall union field is mapped to a Haskell
data constructor.

* Implementend `fieldModifier`

This function controls how a Dhall record field is mapped to a Haskell
record field.

* Fixed: `fieldModifier` in nested types and `defaultGenerateOptions`

* Added tests for `makeHaskellTypesWith`

* Improved documentation

* Fixed: Reference to `InterpretOptions` in documentation

* Use `Dhall.Core.internalError` instead of error

* Fixed: Call to `Dhall.Core.internalError` expects Text not String