Fix dhall-to-nix encoding of nix keys with special symbols
hnix assumes that the parser will do “the right thing” and only use
`StaticKey` for keys that do not contain any special symbols, and the
right `DynamicKey` for everything else.

For example, the dhall

```dhall
{ `text/*` = "foo" }
```

would be encoded as the nix

```nix
{ text/* = "foo"; }
```

which fails the nix parser at `/`.

Instead it should be quoted, like this:

```nix
{ "text/*" = "foo"; }
```

I believe this is a bug in the hnix pretty printer, which should
correctly escape special symbols (for the reference, it does so for
`StaticKey`s that are nix keywords).

But anyway, we can work around this by always constructing a
`DynamicKey` with simple double quotes instead.

I opted against doing some heuristics to encode some “simple” keys
without code, because this should be the task of hnix, not of dhall.