Support GHC 9.4 (#2532)
* chore: enable CPP in Dhall.DirectoryTree

Adding a \ to the escaped newlines appears to make CPP generate valid
haskell without getting stuck.

* feat: support unix-compat 0.8

In unix-compat 0.7, `getUserEntryForName` threw an exception when called
on Windows. unix-compat 0.8 removes it entirely. To increase likelihood
of this change being merged, the behaviour of the 0.7 version has been
incorporated into Dhall.DirectoryTree.

* chore: fix template-haskell bound

dhall does not compile with template-haskell >=2.17 due to signature
change of Language.Haskell.TH.Syntax.PlainTV.

* chore: support template-haskell 2.17 through 2.19

In particular this enables building with GHC 9.4.

* chore(dhall): version 1.42.1

* fix: System.Posix.User.getGroupEntryForName does not exist on Windows