Refactor `dhall-toml` code
This is a (mostly) behavior-preserving refactor of the `dhall-toml`
package since I was planning on working on #2509 and wanted to first
refactor the code a bit more to my liking.

The only actual change is that the `Show` instance for `CompilerError`
is now the derived one and I moved that logic to the `displayException`
method.