Remove `lsp-test` pin
We normally don't run the `dhall-lsp-server` tests by default when
building with Nix, but if you do enable them for building a local
shell for testing purposes then this `lsp-test` pin breaks that shell's
build.  The pin is also not necessary (the default version provided by
Nixpkgs works fine after the upgrade).