to-directory-tree: Fixed --allow-path-separators for fixpointed files
Before, when one uses `to-directory-tree --allow-path-separators` to
create a file using the fixpoint representation of a filesystem tree,
then that failed if a path with path separators was passed to the `file`
or `binary-file` constructors. The reason was that we did not create the
parent directories, which is intended by the use of
--allow-path-separators.

This commit fixes that.