Use unnamed Pi type argument. (#2640)
Now the `FromDhall` and `ToDhall` instances both use unnamed arguments
in their type signature.

The value of this type is not used on type level so it is not necessary
to bring it into scope here.

On the value level this obviously is still necessary but that part is
untouched regarding this commit.

Co-authored-by: gabby <GenuineGabriella@gmail.com>